Move method to shared file
