+++ client/blocks/normalize.js
@@ -9,7 +9,10 @@
  * @typedef {import('@woocommerce/type-defs/billing').BillingData} CartBillingAddress
  */
 
-import { getBlocksConfiguration } from 'wcstripe/blocks/utils';
+import {
+	extractOrderAttributionData,
+	getBlocksConfiguration,
+} from 'wcstripe/blocks/utils';
 
 /**
  * Normalizes order data received upon creating an order using the store's AJAX API.
@@ -80,28 +83,6 @@ const normalizeOrderData = ( paymentMethodEvent, paymentRequestType ) => {
 	}
 
 	return { ...data, ...extractOrderAttributionData() };
-};
-
-/**
- * Get order attribution data from the hidden inputs.
- *
- * @return {Object} Order attribution data.
- */
-export const extractOrderAttributionData = () => {
-	const orderAttributionWrapper = document.getElementsByTagName(
-		'wc-order-attribution-inputs'
-	);
-	if ( ! orderAttributionWrapper.length ) {
-		return {};
-	}
-
-	const orderAttributionData = {};
-	const orderAttributionInputs = orderAttributionWrapper[ 0 ].children;
-	for ( let i = 0; i < orderAttributionInputs.length; i++ ) {
-		orderAttributionData[ orderAttributionInputs[ i ].name ] =
-			orderAttributionInputs[ i ].value;
-	}
-	return orderAttributionData;
 };
 
 /**
